requirements.txt
-i https://pypi.org/simple
aiohappyeyeballs==2.4.3; python_version >= '3.8'
aiohttp==3.10.10; python_version >= '3.8'
aiosignal==1.3.1; python_version >= '3.7'
attrs==24.2.0; python_version >= '3.7'
certifi==2024.8.30; python_version >= '3.6'
charset-normalizer==3.4.0; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
frozenlist==1.5.0; python_version >= '3.8'
gtts==2.5.3; python_version >= '3.7'
idna==3.10; python_version >= '3.6'
iso8601==2.1.0; python_version >= '3.7' and python_version < '4.0'
multidict==6.1.0; python_version >= '3.8'
propcache==0.2.0; python_version >= '3.8'
python-dotenv==1.0.1; python_version >= '3.8'
requests==2.32.3; python_version >= '3.8'
twitchio==2.10.0
typing-extensions==4.12.2; python_version >= '3.8'
urllib3==2.2.3; python_version >= '3.8'
yarl==1.17.1; python_version >= '3.9'
